feat: rename localstorage to localStorageReducer
Renaming all references of localstorage to localStorageReducer
