+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@w3cwebstorage/w3cwebstorage",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "description": "Web Storage with reducers",
   "main": "./dist/index.cjs.js",
   "module": "./dist/index.js",
