+++ src/localStorageReducer/README.md
@@ -17,9 +17,9 @@
 ##### Get a value from local storage
 
 ```js
-import { localstorage } from "@w3cwebstorage/w3cwebstorage";
+import { localStorageReducer } from "@w3cwebstorage/w3cwebstorage";
 
-const value = localstorage(window.localStorage, {
+const value = localStorageReducer(window.localStorage, {
   type: "get",
   key: "key"
 });
@@ -28,9 +28,9 @@ const value = localstorage(window.localStorage, {
 ##### Set a value to local storage
 
 ```js
-import { localstorage } from "@w3cwebstorage/w3cwebstorage";
+import { localStorageReducer } from "@w3cwebstorage/w3cwebstorage";
 
-localstorage(window.localStorage, {
+localStorageReducer(window.localStorage, {
   type: "set",
   key: "key",
   value: "value"
@@ -40,9 +40,9 @@ localstorage(window.localStorage, {
 ##### Remove a value from local storage
 
 ```js
-import { localstorage } from "@w3cwebstorage/w3cwebstorage";
+import { localStorageReducer } from "@w3cwebstorage/w3cwebstorage";
 
-const value = localstorage(window.localStorage, {
+const value = localStorageReducer(window.localStorage, {
   type: "remove",
   key: "key"
 });
@@ -51,7 +51,7 @@ const value = localstorage(window.localStorage, {
 ##### Clear all values in local storage
 
 ```js
-import { localstorage } from "@w3cwebstorage/w3cwebstorage";
+import { localStorageReducer } from "@w3cwebstorage/w3cwebstorage";
 
 const value = localstorage(window.localStorage, {
   type: "clear"
